Move extension to IOperation and add extension methods for property and RaiseEvent.
