+++ analyzers/src/SonarAnalyzer.Common/SymbolicExecution/Roslyn/Extensions/IInvocationOperationExtensions.cs
@@ -42,9 +42,5 @@ namespace SonarAnalyzer.SymbolicExecution.Roslyn
             || (invocation.TargetMethod.IsExtensionMethod
                 && !invocation.Arguments.IsEmpty
                 && state.ResolveCapture(invocation.Arguments[0].ToArgument().Value.UnwrapConversion()).Kind == OperationKindEx.InstanceReference);
-
-        /// <inheritdoc cref="IOperationExtensions.ArgumentValue(ImmutableArray{IOperation}, string)"/>
-        public static IOperation ArgumentValue(this IInvocationOperationWrapper invocation, string parameterName) =>
-            invocation.Arguments.ArgumentValue(parameterName);
     }
 }
